Move CacheEventStore to the final place.
