+++ src/Storage/Caching/SnMemoryCache.cs
@@ -19,7 +19,7 @@ namespace SenseNet.ContentRepository.Storage.Caching
 
         public int Count => (int)base.GetCount();
 
-        public EventStore Events { get; set; }
+        public CacheEventStore Events { get; set; }
 
         public object Get(string key)
         {
