+++ src/Tests/SenseNet.Tests/Implementations/EmptyCache.cs
@@ -11,7 +11,7 @@ namespace SenseNet.Tests.Implementations
         // ReSharper disable once CollectionNeverUpdated.Local
         private readonly Dictionary<string, object> _emptyCache = new Dictionary<string, object>();
 
-        public EventStore Events { get; set; }
+        public CacheEventStore Events { get; set; }
 
         public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
         {
